Add support for watching file descriptors in glib.
This is useful for interacting with the GTK main loop from another
goroutine. GTK has to be blocking in the glib main loop in order to
process UI events so other goroutines have to poke it via a means which
is compatible with that and a socket/pipe works nicely.