Alter glib.List to return unsafe.Pointers
Previously, glib.List.Data and NthData would return a cgo gpointer
wrapped in an interface{}. This change alters that to be an
unsafe.Pointer wrapped in an interface{} so that the underlying
type isn't private.

It also alters GetChildren to return a list of widgets, rather than
a glib.List.